New File for teacher class
