+++ ASI.Basecode.WebApp/Views/ClassManagement/ViewClass.cshtml
@@ -6,7 +6,7 @@
 }
 
 <div class="container mt-4">
-    <a asp-action="TeacherDashboard" class="text-decoration-none text-muted mb-3 d-inline-block">
+    <a asp-controller="TeacherClass" asp-action="TeacherDashboard" class="text-decoration-none text-muted mb-3 d-inline-block">
         <i class="fa-solid fa-arrow-left me-1"></i> Back to Dashboard
     </a>
