+++ ASI.Basecode.WebApp/Views/Shared/_TeacherLayout.cshtml
@@ -62,7 +62,7 @@
             <nav class="mt-4 px-2">
                 
                 <div class="mb-2">
-                    <a asp-controller="ClassManagement" asp-action="TeacherDashboard"
+                    <a asp-controller="TeacherClass" asp-action="TeacherDashboard"
                        class="flex items-center px-4 py-3 text-sm font-medium text-gray-700 hover:bg-gray-100 cursor-pointer sidebar-item transition-colors duration-200 rounded-lg">
                         <i class="fas fa-tachometer-alt mr-3 text-gray-500"></i>
                         Dashboard
