+++ ASI.Basecode.WebApp/Views/TeacherGrade/ClassGrades.cshtml
@@ -9,7 +9,7 @@
 
 <div class="container mx-auto px-4 py-6">
     <div class="mb-3">
-        <a asp-controller="ClassManagement" asp-action="TeacherDashboard" class="inline-flex items-center px-3 py-2 border rounded bg-white text-sm text-gray-700 mb-3">
+        <a asp-controller="TeacherClass" asp-action="TeacherDashboard" class="inline-flex items-center px-3 py-2 border rounded bg-white text-sm text-gray-700 mb-3">
             <i class="fas fa-chevron-left me-2"></i> Back to Dashboard
         </a>
     </div>
